requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.6.2.post1
fastapi==0.115.5
uvicorn
idna==3.10
Mako==1.3.6
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
pydantic==2.10.1
pydantic_core==2.27.1
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.36
SQLAlchemy-Utils==0.41.2
starlette==0.41.3
typing_extensions==4.12.2
Flask==2.3.3